+++ maroonxdr/maroonx/tests/echelle_extraction/test_dark_subtraction.py
@@ -8,6 +8,9 @@ from maroonxdr.maroonx.primitives_maroonx_echelle import MAROONXEchelle
 
 @pytest.mark.parametrize("filename", ["./maroonxdr/maroonx/tests/echelle_extraction/20220808T111549Z_SOOOE_r_0300_reduced.fits"])
 def test_getting_stripe_locations(caplog, filename):
+    """
+    This test checks that the stripe locations for dark subtraction are being found correctly.
+    """
     caplog.set_level(logging.DEBUG)
     ad = astrodata.open(filename)
     p = MAROONXEchelle([deepcopy(ad)])